# internal/controller/amqp_rpc/v1/response/task.go
package response

import "github.com/alfariesh/backend-memora/internal/entity"

// TaskList -.
type TaskList struct {
	Tasks []entity.Task `json:"tasks"`
	Total int           `json:"total"`
}

// DeleteStatus -.
type DeleteStatus struct {
	Status string `json:"status"`
}